+++ packages/dsw-document-worker/CHANGELOG.md
@@ -8,6 +8,17 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
 
 
+## [4.20.0]
+
+### Added
+
+- Support item title for all types of questions
+- Add Jinja filter and mechanism to extract replies using KM annotations 
+
+### Fixed
+
+- Fixed file replies linked from file questions
+
 ## [4.19.1]
 
 ### Changed
@@ -440,3 +451,4 @@ Released for version consistency with other DSW tools.
 [4.18.3]: /../../tree/v4.18.3
 [4.19.0]: /../../tree/v4.19.0
 [4.19.1]: /../../tree/v4.19.1
+[4.20.0]: /../../tree/v4.20.0